requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.20.0
accelerate>=0.20.0

# Data processing and machine learning
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.0.0

# Visualization and monitoring
matplotlib>=3.5.0
tensorboard>=2.10.0

# Image processing
opencv-python>=4.5.0

# GUI and interactive features (for interactive_diff_generator.py)
tkinter  # Usually comes with Python, but listed for clarity

# Text processing and utilities
textwrap  # Built-in Python module

# Optimal transport (optional, for OT-based alignment)
# geomloss>=0.2.0

# Other utilities
tqdm>=4.60.0
argparse  # Built-in Python module
json5>=0.9.0

# Development tools (optional)
# jupyter>=1.0.0
# ipython>=7.0.0 